+++ lib/config/router.dart
@@ -1,12 +1,12 @@
-import 'package:danuri_flutter/view/failure/screen/failure_screen.dart';
-import 'package:danuri_flutter/view/completion/screen/completion_screen.dart';
-import 'package:danuri_flutter/view/item_rental/screen/item_rental_screen.dart';
-import 'package:danuri_flutter/view/register_used_space/screen/register_used_space.dart';
-import 'package:danuri_flutter/view/auth_code_login/screen/auth_code_login_screen.dart';
-import 'package:danuri_flutter/view/sign_up/screen/sign_up_screen.dart';
-import 'package:danuri_flutter/view/login/screen/login_screen.dart';
-import 'package:danuri_flutter/view/organ_auth/screen/organ_auth_screen.dart';
-import 'package:danuri_flutter/view/home/screen/home_screen.dart';
+import 'package:danuri_flutter/presentation/failure/screen/failure_screen.dart';
+import 'package:danuri_flutter/presentation/completion/screen/completion_screen.dart';
+import 'package:danuri_flutter/presentation/item_rental/screen/item_rental_screen.dart';
+import 'package:danuri_flutter/presentation/register_used_space/screen/register_used_space.dart';
+import 'package:danuri_flutter/presentation/auth_code_login/screen/auth_code_login_screen.dart';
+import 'package:danuri_flutter/presentation/sign_up/screen/sign_up_screen.dart';
+import 'package:danuri_flutter/presentation/login/screen/login_screen.dart';
+import 'package:danuri_flutter/presentation/organ_auth/screen/organ_auth_screen.dart';
+import 'package:danuri_flutter/presentation/home/screen/home_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
