+++ lib/presentation/widgets/button/help_me_button.dart
@@ -1,0 +1,36 @@
+import 'package:danuri_flutter/core/theme/color.dart';
+import 'package:danuri_flutter/core/theme/text.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class HelpMeButton extends StatelessWidget {
+  const HelpMeButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        //미정
+      },
+      child: Container(
+        width: 152.w,
+        height: 48.h,
+        decoration: ShapeDecoration(
+          color: DanuriColor.line1,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              '? 도와주세요',
+              style: DanuriText.body1Normal,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
